Add infrastructure for our own global sbt dir
Okay, this commit adds infrastructure so that bloop can inject logic to
every source dependency it depends on. It is basically a transparent
workaround to all the limitations that sbt imposes on us.

This way, we can control remove all the crazy stuff we're doing on every
state reload, and we can programmatically add synthetic projects and
derive projects of every source dependency build (because of the Plugin
API). That includes enabling and disabling compiler plugins.

The best thing of all is that users of this build don't need to care
about anything. The build will set up itself automatically the first
time the repository is cloned, and then it will exit and ask the user to
run sbt again, this time with all the things set up.